pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "um-agent-coder"
version = "0.2.0"
description = "Multi-model AI coding agent with parallel execution, task decomposition, and orchestration for long-running tasks"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "UMwai"}
]
keywords = [
    "ai",
    "agent",
    "llm",
    "coding",
    "orchestration",
    "multi-model",
    "parallel-execution",
    "task-decomposition"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31.0",
    "typing-extensions>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.0",
    "ruff>=0.1.0",
]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.18.0",
]
google = [
    "google-generativeai>=0.3.0",
]
all = [
    "um-agent-coder[openai,anthropic,google]",
]
daemon = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "aiosqlite>=0.19.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.26.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "google-generativeai>=0.8.0",
    "google-auth>=2.27.0",
    "google-cloud-secret-manager>=2.16.0",
    "google-cloud-firestore>=2.16.0",
]

[project.urls]
Homepage = "https://github.com/UMwai/um-agent-coder"
Repository = "https://github.com/UMwai/um-agent-coder"
Issues = "https://github.com/UMwai/um-agent-coder/issues"

[project.scripts]
um-agent = "um_agent_coder.cli:main"
um-agent-coder = "um_agent_coder.cli:main"
um-agent-daemon = "um_agent_coder.daemon.app:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["um_agent_coder*"]

[tool.setuptools.package-data]
um_agent_coder = ["py.typed", "daemon/static/*"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501", "UP006", "UP007", "UP035", "UP045", "N806", "B904"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
